refactor translations
Translations (all json) files are now searched by there used schema. The language is now in the json itself, the filename no longer matters.

Translation of scripts are specialy handled. The fields are not the to translated text, but the title creator and field name seperated by `.`

There is now also a schema for translation files (required to search for them) that will warn you if you have translations that are no longer used (hinting at a changed key)

related #23